Add an exemption for archiving cloud-gov repos
Archiving stale 18F repos makes tons of sense, because we don't want dependency alerting for projects that are not deployed and/or maintained. 

Since cloud.gov is a mature product, we don't have the same needs. For security and compliance, we want to make sure we are getting dependency update alerts.